+++ packages/messaging/src/index.ts
@@ -3,3 +3,4 @@ export { EventMessagingType } from './events/index.js';
 export type { EventRegistry } from './events/index.js';
 export * from './jobs/index.js';
 export * from './websockets/index.js';
+export * from './helpers/index.js';